+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## v29.0.0-SNAPSHOT - under development
 
+### 💥 Breaking Changes
+
+* The GridModel `contextMenuFn` parameter has been replaced with a `contextMenu` parameter.  The new
+parameter will allow context menus to be specifed with a simple array in addition to the function
+specification currently supported.
+* The GridModel `defaultContextMenuTokens` array has been renamed `defaultContextMenu`.
+
 ### 🐞 Bug Fixes
 
 * Fixed autoFocus on NumberInput.